Adição da Tela Principal

+++ src/main/java/com/project/DTO/ConfigDTO.java
@@ -17,16 +17,37 @@ public class ConfigDTO {
         try {
             System.out.println("Digite o caminho para criar a pasta onde ficará os livros");
             setCaminho(new Scanner(System.in).next() + "/library");
-            System.out.println(getCaminho());
             File pasta = new File(getCaminho());
-            pasta.mkdir();
-            pasta = new File(getCaminho() + "/config.txt");
-            pasta.createNewFile();
-            try (BufferedWriter escrever = new BufferedWriter(new FileWriter(caminho + "/config.txt", false))) {
-                escrever.write("Livros:");
-            } catch (Exception e) {
-                System.out.println("Erro " + e);
+
+            if (pasta.exists() && pasta.isDirectory()) {
+                System.out.println("Repositório já existe");
+                pasta = new File(getCaminho() + "/config.txt");
+                if (pasta.exists()) {
+                    System.out.println("já está tudo previamente pronto para o uso");
+                } else {
+                    pasta.createNewFile();
+                    try (BufferedWriter escrever = new BufferedWriter(new FileWriter(caminho + "/config.txt", false))) {
+                        escrever.write("Livros:");
+                    } catch (Exception e) {
+                        System.out.println("Erro " + e);
+                    }
+                }
+
+            } else {
+                pasta.mkdir();
+                pasta = new File(getCaminho() + "/config.txt");
+                if (pasta.exists()) {
+                    System.out.println("já está tudo previamente pronto para o uso");
+                } else {
+                    pasta.createNewFile();
+                    try (BufferedWriter escrever = new BufferedWriter(new FileWriter(caminho + "/config.txt", false))) {
+                        escrever.write("Livros:");
+                    } catch (Exception e) {
+                        System.out.println("Erro " + e);
+                    }
+                }
             }
+
         } catch (IOException e) {
             System.out.println("Erro: " + e);
         }
@@ -46,16 +67,16 @@ public class ConfigDTO {
                     System.out.println(livro.getName());
                 }
             }
-             System.out.println("Os " + livros.size() + " livros que foram encontrados são: ");
+            organizarTela();
+            System.out.println("Os " + livros.size() + " livros que foram encontrados são: ");
             for (String l : livros) {
-                if(!checarLivros(l)){
+                if (!checarLivros(l)) {
                     inserirLivro(l);
-                     System.out.println(l);
-            }
-               
+                    System.out.println(l);
+                }
 
             }
-
+            organizarTela();
         }
 
         System.out.println("Caso um livro não esteja na lista verifique que ele está em formato .pdf");
@@ -66,9 +87,9 @@ public class ConfigDTO {
         try (BufferedReader read = new BufferedReader(new FileReader(getCaminho() + "/config.txt"))) {
             String s;
             while ((s = read.readLine()) != null) {
-                    if((s.replace("=0", "")).equals(nome)){
+                if ((s.split("=")[0]).equals(nome.split("=")[0])) {
                     return true;
-                    }
+                }
             }
 
         } catch (IOException e) {
@@ -76,23 +97,29 @@ public class ConfigDTO {
         }
         return false;
     }
-    
-    
-    
-    
-    
-    public void inserirLivro(String nome){
-        try (BufferedWriter escrever = new BufferedWriter(new FileWriter(getCaminho()+"/config.txt",true))){
-            escrever.write("\n"+nome+"=0");
-            
-            
-            
+
+    public void inserirLivro(String nome) {
+        try (BufferedWriter escrever = new BufferedWriter(new FileWriter(getCaminho() + "/config.txt", true))) {
+            escrever.write("\n" + nome + "=0");
+
         } catch (IOException e) {
         }
-        
+
     }
-  
     
+     public void listaLivros() {
+
+        String line;
+        try (BufferedReader read = new BufferedReader(new FileReader(getCaminho() + "/config.txt"))) {
+            while ((line = read.readLine()) != null) {
+                System.out.println(line.split(".pdf")[0]);
+            }
+        } catch (IOException e) {
+            System.out.println("Erro: " + e);
+        }
+        organizarTela();
+    }
+
     public String getCaminho() {
         return caminho;
     }
@@ -101,4 +128,16 @@ public class ConfigDTO {
         this.caminho = caminho;
     }
 
+    public void organizarTela() {
+        for (int i = 0; i < 3; i++) {
+            System.out.println(" ");
+        }
+        for (int i = 0; i < 20; i++) {
+            System.out.print("-");
+        }
+        for (int i = 0; i < 3; i++) {
+            System.out.println(" ");
+        }
+
+    }
 }
